Change the structure of the project and tests.
