+++ kinto_updater/exceptions.py
@@ -1,0 +1,2 @@
+class UpdaterException(Exception):
+    pass
