+++ kinto_updater/signer.py
