+++ kinto_updater/tests/test_signer.py
