+++ kinto_updater/updater.py
@@ -1,0 +1,71 @@
+import copy
+import uuid
+
+from kinto_client import Client
+
+import hasher
+import signer
+import exceptions
+
+
+class Updater(object):
+
+    def __init__(self, bucket, collection, server_url=None,
+                 auth=None, session=None, endpoints=None,
+                 signer_instance=None, settings=None):
+        self.client = Client(
+            bucket=bucket, collection=collection,
+            server_url=server_url, session=session)
+
+        if settings is None:
+            settings = {}
+        self.settings = settings
+
+        if signer_instance is None:
+            signer_instance = signer.RSABackend(self.settings)
+        self.signer_instance = signer_instance
+
+        self.bucket = bucket
+        self.collection = collection
+
+    def gather_remote_collection(self):
+        '''Retrieves the remote collection and returns it.'''
+        collection = self.client.get_collection()
+        records = {record['id']: record
+                   for record in self.client.get_records()}
+
+        return records, collection['data']
+
+    def check_data_validity(self, records, signature):
+        local_hash = hasher.compute_hash(records.values())
+        self.signer_instance.verify(local_hash, signature)
+
+    def add_records(self, new_records):
+        new_records = copy.deepcopy(new_records)
+        records, collection_data = self.gather_remote_collection()
+
+        if records:
+            if 'signature' not in collection_data:
+                msg = "Unable to verify unsigned data"
+                raise exceptions.UpdaterException(msg)
+            signature = collection_data['signature']
+            self.check_data_validity(records, signature)
+
+        with self.client.batch() as batch:
+            # Create IDs for records which don't already have one.
+            for record in new_records:
+
+                if 'id' not in record:
+                    record['id'] = str(uuid.uuid4())
+                    batch.create_record(id=record['id'], data=record)
+                else:
+                    batch.update_record(id=record['id'], data=record)
+
+            # Compute the hash of the old + new records
+            records.update({record['id']: record
+                            for record in new_records})
+            new_hash = hasher.compute_hash(records.values())
+            signature = self.signer_instance.sign(new_hash)
+
+            # Send the new hash + signature to the remote.
+            batch.patch_collection(data={'signature': signature})
